Extract folded model drawing.
