+++ src/main/java/oripa/renderer/estimation/CoordinateConverter.java
@@ -16,7 +16,7 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package oripa.swing.view.estimation;
+package oripa.renderer.estimation;
 
 import oripa.geom.RectangleDomain;
 import oripa.vecmath.Vector2d;
@@ -26,7 +26,7 @@ import oripa.vecmath.Vector3d;
  * @author OUCHI Koji
  *
  */
-class CoordinateConverter {
+public class CoordinateConverter {
 	private final double imageHeight;
 	private final double imageWidth;
