+++ src/main/java/oripa/renderer/estimation/DistortionMethod.java
@@ -1,6 +1,7 @@
 /**
- * ORIPA - Origami Pattern Editor
- * Copyright (C) 2005-2009 Jun Mitani http://mitani.cs.tsukuba.ac.jp/
+ * ORIPA - Origami Pattern Editor 
+ * Copyright (C) 2013-     ORIPA OSS Project  https://github.com/oripa/oripa
+ * Copyright (C) 2005-2009 Jun Mitani         http://mitani.cs.tsukuba.ac.jp/
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -15,16 +16,10 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+package oripa.renderer.estimation;
 
-package oripa.swing.view.estimation;
-
-import oripa.vecmath.Vector2d;
-
-public class TriangleVertex {
-
-	// position after folding
-	public Vector2d p = new Vector2d(0, 0);
-
-	public Vector2d uv = new Vector2d(0, 0);
-	public FloatingRGB color = new FloatingRGB(0, 0, 0);
+public enum DistortionMethod {
+	NONE,
+	DEPTH,
+	MORISUE
 }
