+++ src/main/java/oripa/renderer/estimation/OverlapRelationInterpolater.java
@@ -16,7 +16,7 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package oripa.swing.view.estimation;
+package oripa.renderer.estimation;
 
 import java.util.List;
 
@@ -27,7 +27,7 @@ import oripa.domain.fold.origeom.OverlapRelation;
  * @author OUCHI Koji
  *
  */
-class OverlapRelationInterpolater {
+public class OverlapRelationInterpolater {
 	public OverlapRelation interpolate(final OverlapRelation overlapRelation, final List<Face> faces,
 			final double eps) {
 		var interpolatedOverlapRelation = overlapRelation.clone();
