+++ src/main/java/oripa/renderer/estimation/TriangleVertex.java
@@ -1,7 +1,6 @@
 /**
  * ORIPA - Origami Pattern Editor
- * Copyright (C) 2013-     ORIPA OSS Project  https://github.com/oripa/oripa
- * Copyright (C) 2005-2009 Jun Mitani         http://mitani.cs.tsukuba.ac.jp/
+ * Copyright (C) 2005-2009 Jun Mitani http://mitani.cs.tsukuba.ac.jp/
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -16,37 +15,16 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package oripa.swing.view.estimation;
 
-/**
- * @author OUCHI Koji
- *
- */
-public class FloatingRGB {
-	private final double r, g, b;
+package oripa.renderer.estimation;
 
-	public FloatingRGB(final double r, final double g, final double b) {
-		this.r = r;
-		this.g = g;
-		this.b = b;
-	}
+import oripa.vecmath.Vector2d;
 
-	public FloatingRGB(final FloatingRGB rgb) {
-		r = rgb.r;
-		g = rgb.g;
-		b = rgb.b;
-	}
+public class TriangleVertex {
 
-	public double getR() {
-		return r;
-	}
+	// position after folding
+	public Vector2d p = new Vector2d(0, 0);
 
-	public double getG() {
-		return g;
-	}
-
-	public double getB() {
-		return b;
-	}
-
+	public Vector2d uv = new Vector2d(0, 0);
+	public FloatingRGB color = new FloatingRGB(0, 0, 0);
 }
