+++ src/main/java/oripa/renderer/estimation/VertexColorMapFactory.java
@@ -16,7 +16,7 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package oripa.swing.view.estimation;
+package oripa.renderer.estimation;
 
 import java.util.HashMap;
 import java.util.List;
