+++ src/main/java/oripa/renderer/estimation/VertexDepthMapFactory.java
@@ -16,7 +16,7 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package oripa.swing.view.estimation;
+package oripa.renderer.estimation;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -35,7 +35,7 @@ import oripa.value.OriPoint;
  * @author OUCHI Koji
  *
  */
-class VertexDepthMapFactory {
+public class VertexDepthMapFactory {
 	public Map<OriVertex, Integer> create(final OrigamiModel origamiModel, final OverlapRelation overlapRelation,
 			final double eps) {
