+++ src/main/java/oripa/swing/view/estimation/FoldedModelScreen.java
@@ -22,7 +22,6 @@ import java.awt.AlphaComposite;
 import java.awt.Color;
 import java.awt.Graphics;
 import java.awt.Graphics2D;
-import java.awt.Image;
 import java.awt.RenderingHints;
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
@@ -32,12 +31,8 @@ import java.awt.event.MouseWheelListener;
 import java.awt.geom.AffineTransform;
 import java.awt.geom.Point2D;
 import java.awt.image.BufferedImage;
-import java.awt.image.MemoryImageSource;
-import java.io.File;
-import java.util.List;
 import java.util.Map;
 
-import javax.imageio.ImageIO;
 import javax.swing.JPanel;
 
 import oripa.domain.fold.FoldedModel;
@@ -45,10 +40,14 @@ import oripa.domain.fold.halfedge.OriFace;
 import oripa.domain.fold.halfedge.OriVertex;
 import oripa.domain.fold.halfedge.OrigamiModel;
 import oripa.domain.fold.origeom.OverlapRelation;
-import oripa.domain.fold.origeom.OverlapRelationValues;
 import oripa.geom.RectangleDomain;
 import oripa.gui.view.estimation.DefaultColors;
+import oripa.renderer.estimation.DistortionFacade;
+import oripa.renderer.estimation.DistortionMethod;
+import oripa.renderer.estimation.FoldedModelPixelRenderer;
+import oripa.renderer.estimation.VertexDepthMapFactory;
 import oripa.swing.drawer.java2d.GraphicItemConverter;
+import oripa.swing.drawer.java2d.PixelDrawer;
 import oripa.swing.view.util.MouseUtility;
 import oripa.vecmath.Vector2d;
 
@@ -61,39 +60,21 @@ import oripa.vecmath.Vector2d;
 public class FoldedModelScreen extends JPanel
 		implements MouseListener, MouseMotionListener, MouseWheelListener {
 
-	private BufferedImage bufferImage;
-	private Graphics2D bufferg;
-	private final int pbuf[]; // 32bit pixel buffer
-	private final int zbuf[]; // 32bit z buffer
-	private final int BUFFERW; // width
-	private final int BUFFERH; // height
-	private final int min[];
-	private final int max[];
-	private final int minr[];
-	private final int maxr[];
-	private final int ming[];
-	private final int maxg[];
-	private final int minb[];
-	private final int maxb[];
-	private final double minu[];
-	private final double maxu[];
-	private final double minv[];
-	private final double maxv[];
+	private final BufferedImage bufferImage;
+	private static final int BUFFERW = 600; // width
+	private static final int BUFFERH = 600; // height
 	private boolean useColor = true;
 	private boolean fillFaces = true;
 	private boolean ambientOcclusion = false;
 	private boolean faceOrderFlip = false;
 	private final double scaleRate = 0.8;
 	private boolean drawEdges = true;
-	private Image renderImage;
 	private double rotateAngle;
 	private double scale;
 	private double transX;
 	private double transY;
 	private Point2D preMousePoint;
 	private final AffineTransform affineTransform;
-	private BufferedImage textureImage = null;
-	private final boolean bUseTexture = false;
 
 	private Color frontColor = DefaultColors.FRONT;
 	private Color backColor = DefaultColors.BACK;
@@ -101,8 +82,9 @@ public class FoldedModelScreen extends JPanel
 
 	private OrigamiModel origamiModel = null;
 	private OverlapRelation overlapRelation;
-	private OverlapRelation interpolatedOverlapRelation;
 	private OriFace selectedSubface = null;
+
+	FoldedModelPixelRenderer pixelRenderer;
 
 	private RectangleDomain domain;
 
@@ -110,6 +92,7 @@ public class FoldedModelScreen extends JPanel
 	private double eps = 1e-5;
 
 	private DistortionMethod distortionMethod = DistortionMethod.NONE;
+	private DistortionFacade distortion;
 
 	private Map<OriVertex, Integer> vertexDepths;
 
@@ -118,39 +101,13 @@ public class FoldedModelScreen extends JPanel
 		addMouseMotionListener(this);
 		addMouseWheelListener(this);
 
-		BUFFERW = 600;
-		BUFFERH = 600;
+		pixelRenderer = new FoldedModelPixelRenderer(BUFFERW, BUFFERH);
 
-		pbuf = new int[BUFFERW * BUFFERH];
-		zbuf = new int[BUFFERW * BUFFERH];
-		min = new int[BUFFERH];
-		max = new int[BUFFERH];
-		minr = new int[BUFFERH];
-		maxr = new int[BUFFERH];
-		ming = new int[BUFFERH];
-		maxg = new int[BUFFERH];
-		minb = new int[BUFFERH];
-		maxb = new int[BUFFERH];
-		maxu = new double[BUFFERH];
-		maxv = new double[BUFFERH];
-		minu = new double[BUFFERH];
-		minv = new double[BUFFERH];
-
-		clear();
-		drawOrigami();
 		rotateAngle = 0;
 		scale = 1.0;
 		affineTransform = new AffineTransform();
-		updateAffineTransform();
 
-		if (bUseTexture) {
-			try {
-				textureImage = ImageIO.read(new File("c:\\chiyo2-1024.bmp"));
-			} catch (Exception e) {
-				e.printStackTrace();
-				textureImage = null;
-			}
-		}
+		bufferImage = new BufferedImage(pixelRenderer.width, pixelRenderer.height, BufferedImage.TYPE_INT_RGB);
 	}
 
 	private void resetViewMatrix() {
@@ -162,13 +119,13 @@ public class FoldedModelScreen extends JPanel
 		}
 
 		domain = origamiModel.createDomainOfFoldedModel();
+		distortion = new DistortionFacade(domain, BUFFERW, BUFFERH);
 
 		updateAffineTransform();
 		redrawOrigami();
 	}
 
 	public void redrawOrigami() {
-		clear();
 		drawOrigami();
 		repaint();
 	}
@@ -214,105 +171,14 @@ public class FoldedModelScreen extends JPanel
 		redrawOrigami();
 	}
 
-	private int getIndex(final int x, final int y) {
-		return y * BUFFERW + x;
-	}
-
-	public void clear() {
-		for (int i = 0; i < BUFFERW * BUFFERH; i++) {
-			pbuf[i] = 0xffffffff;
-			zbuf[i] = -1;
-		}
-	}
-
 	private void updateAffineTransform() {
 		affineTransform.setToIdentity();
-		affineTransform.translate(getWidth() * 0.5, getHeight() * 0.5);
+		affineTransform.translate(pixelRenderer.width * 0.5, pixelRenderer.height * 0.5);
 		affineTransform.scale(scale, scale);
 		affineTransform.translate(transX, transY);
 		affineTransform.rotate(rotateAngle);
-		affineTransform.translate(-getWidth() * 0.5, -getHeight() * 0.5);
+		affineTransform.translate(-pixelRenderer.width * 0.5, -pixelRenderer.height * 0.5);
 	}
-
-	/*
-	 * Convenience method that returns a scaled instance of the provided {@code
-	 * BufferedImage}.
-	 *
-	 * @param img the original image to be scaled
-	 *
-	 * @param targetWidth the desired width of the scaled instance, in pixels
-	 *
-	 * @param targetHeight the desired height of the scaled instance, in pixels
-	 *
-	 * @param hint one of the rendering hints that corresponds to {@code
-	 * RenderingHints.KEY_INTERPOLATION} (e.g. {@code
-	 * RenderingHints.VALUE_INTERPOLATION_NEAREST_NEIGHBOR}, {@code
-	 * RenderingHints.VALUE_INTERPOLATION_BILINEAR}, {@code
-	 * RenderingHints.VALUE_INTERPOLATION_BICUBIC})
-	 *
-	 * @param higherQuality if true, this method will use a multi-step scaling
-	 * technique that provides higher quality than the usual one-step technique
-	 * (only useful in downscaling cases, where {@code targetWidth} or {@code
-	 * targetHeight} is smaller than the original dimensions, and generally only
-	 * when the {@code BILINEAR} hint is specified)
-	 *
-	 * @return a scaled version of the original {@code BufferedImage}
-	 */
-//	public BufferedImage getScaledInstance(final BufferedImage img,
-//			final int targetWidth,
-//			final int targetHeight,
-//			final Object hint,
-//			final boolean higherQuality) {
-//		int type = (img.getTransparency() == Transparency.OPAQUE)
-//				? BufferedImage.TYPE_INT_RGB
-//				: BufferedImage.TYPE_INT_ARGB;
-//		BufferedImage ret = img;
-//		int w, h;
-//		if (higherQuality) {
-//			// Use multi-step technique: start with original size, then
-//			// scale down in multiple passes with drawImage()
-//			// until the target size is reached
-//			w = img.getWidth();
-//			h = img.getHeight();
-//			if (w < targetWidth) {
-//				w = targetWidth;
-//			}
-//			if (h < targetHeight) {
-//				h = targetHeight;
-//			}
-//		} else {
-//			// Use one-step technique: scale directly from original
-//			// size to target size with a single drawImage() call
-//			w = targetWidth;
-//			h = targetHeight;
-//		}
-//
-//		do {
-//			if (higherQuality && w > targetWidth) {
-//				w /= 2;
-//				if (w < targetWidth) {
-//					w = targetWidth;
-//				}
-//			}
-//
-//			if (higherQuality && h > targetHeight) {
-//				h /= 2;
-//				if (h < targetHeight) {
-//					h = targetHeight;
-//				}
-//			}
-//
-//			BufferedImage tmp = new BufferedImage(w, h, type);
-//			Graphics2D g2 = tmp.createGraphics();
-//			g2.setRenderingHint(RenderingHints.KEY_INTERPOLATION, hint);
-//			g2.drawImage(ret, 0, 0, w, h, null);
-//			g2.dispose();
-//
-//			ret = tmp;
-//		} while (w != targetWidth || h != targetHeight);
-//
-//		return ret;
-//	}
 
 	public void setModel(final FoldedModel foldedModel, final int overlapRelationIndex, final double eps) {
 		this.eps = eps;
@@ -365,84 +231,64 @@ public class FoldedModelScreen extends JPanel
 	public void paintComponent(final Graphics g) {
 		super.paintComponent(g);
 
-		if (bufferImage == null) {
-			bufferImage = new BufferedImage(getWidth(), getHeight(), BufferedImage.TYPE_INT_RGB);
-			bufferg = (Graphics2D) bufferImage.getGraphics();
-
-			updateAffineTransform();
+		if (pixelRenderer == null) {
+			return;
 		}
+
+		var bufferg = bufferImage.createGraphics();
+
 		bufferg.setRenderingHint(RenderingHints.KEY_ANTIALIASING,
 				RenderingHints.VALUE_ANTIALIAS_ON);
 
 		bufferg.setTransform(new AffineTransform());
-
-		// Clear image
+//		// Clear image
 		bufferg.setComposite(AlphaComposite.getInstance(AlphaComposite.SRC_OVER, 1.0f));
 		bufferg.setColor(Color.WHITE);
 		bufferg.fillRect(0, 0, getWidth(), getHeight());
 
 		bufferg.setTransform(affineTransform);
 
-		Graphics2D g2d = bufferg;
+		var drawer = new PixelDrawer();
 
-		if (renderImage != null) {
-			g2d.drawImage(renderImage, 0, 0, null);
-		}
+		drawer.draw(bufferg, pixelRenderer.getPixels(), pixelRenderer.width, pixelRenderer.height);
+//		bufferg.drawImage(modelImage, 0, 0, null);
 
 		if (selectedSubface != null) {
-			drawSubface(g2d);
+			drawSubface(bufferg);
 		}
 
 		g.drawImage(bufferImage, 0, 0, this);
-	}
 
-	private CoordinateConverter createCoordinateConverter() {
-		var converter = new CoordinateConverter(domain, BUFFERW, BUFFERH);
-
-		converter.setDistortionMethod(distortionMethod);
-		converter.setDistortionParameter(distortionParameter);
-		converter.setScale(getFinalScale());
-
-		return converter;
 	}
 
 	public void drawOrigami() {
 		if (origamiModel == null || overlapRelation == null) {
-			renderImage = createImage(new MemoryImageSource(BUFFERW, BUFFERH, pbuf, 0, BUFFERW));
 			return;
 		}
 
 		if (!origamiModel.isFolded()) {
 			return;
 		}
-		long time0 = System.currentTimeMillis();
 
-		var factory = new FaceFactory(createCoordinateConverter(), vertexDepths);
-		var faces = origamiModel.getFaces().stream()
-				.map(factory::create)
-				.toList();
+		var converter = distortion.createCoordinateConverter(distortionMethod, distortionParameter, getFinalScale());
 
-		interpolatedOverlapRelation = new OverlapRelationInterpolater().interpolate(overlapRelation, faces, eps);
+		var distortionResult = distortion.apply(origamiModel, overlapRelation, converter, vertexDepths, eps);
 
-		faces.forEach(this::drawFace);
+		var faces = distortionResult.getFaces();
+		var interpolatedOverlapRelation = distortionResult.getInterpolatedOverlapRelation();
 
-		if (drawEdges) {
-			drawEdges();
-		}
-
-		if (ambientOcclusion) {
-			applyAmbientOcculusion();
-		}
-
-		long time1 = System.currentTimeMillis();
-
-		// System.out.println("render time = " + (time1 - time0) + "ms");
-
-		renderImage = createImage(new MemoryImageSource(BUFFERW, BUFFERH, pbuf, 0, BUFFERW));
+		pixelRenderer.render(faces, interpolatedOverlapRelation, origamiModel.createPaperDomain(),
+				new FoldedModelPixelRenderer.Option()
+						.setAmbientOcclusion(ambientOcclusion)
+						.setDrawEdges(drawEdges)
+						.setFaceOrderFlipped(faceOrderFlip)
+						.setFillFace(fillFaces)
+						.setUseColor(useColor)
+						.setColors(frontColor, backColor, singleColor));
 	}
 
 	private void drawSubface(final Graphics2D g2d) {
-		var converter = createCoordinateConverter();
+		var converter = distortion.createCoordinateConverter(distortionMethod, distortionParameter, getFinalScale());
 		var convertedSubface = selectedSubface.halfedgeStream()
 				.map(v -> converter.convert(v.getPosition(), 0, v.getPositionBeforeFolding()))
 				.toList();
@@ -453,298 +299,10 @@ public class FoldedModelScreen extends JPanel
 		g2d.draw(path2d);
 	}
 
-	private void drawFace(final Face face) {
-		if (useColor) {
-			drawFace(face, frontColor, backColor);
-		} else {
-			drawFace(face, singleColor, singleColor);
-		}
-	}
-
-	private void drawFace(final Face face, final Color frontColor, final Color backColor) {
-
-		var paperDomain = origamiModel.createPaperDomain();
-
-		List<Double> frontColorFactor;
-		List<Double> backColorFactor;
-
-		frontColorFactor = createColorFactor(frontColor);
-		backColorFactor = createColorFactor(backColor);
-
-		var vertexColorMapFactory = new VertexColorMapFactory();
-		var colorMap = vertexColorMapFactory.createVertexColors(
-				face.getOriginalFace(),
-				frontColorFactor,
-				backColorFactor,
-				isFaceOrderFlipped());
-
-		var triangleFactory = new TriangleFaceFactory();
-		var triangles = triangleFactory.createAll(face);
-		triangles.forEach(triangle -> triangle.prepareColor(colorMap, paperDomain));
-
-		triangles.stream().forEach(tri -> {
-			drawTriangle(tri, face.getFaceID());
-		});
-
-	}
-
 	private double getFinalScale() {
 		return scaleRate * Math.min(
-				BUFFERW / (domain.getWidth()),
-				BUFFERH / (domain.getHeight())) * 0.95;
-	}
-
-	private List<Double> createColorFactor(final Color color) {
-		return List.of(color.getRed(), color.getGreen(), color.getBlue()).stream()
-				.map(c -> c / 255.0)
-				.toList();
-	}
-
-	// --------------------------------------------------------------------
-	// Polygon drawing
-	//
-	// --------------------------------------------------------------------
-	private void drawTriangle(final TriangleFace tri, final int id) {
-
-		// (For speed) set the range of use of the buffer
-		int top = Integer.MAX_VALUE; // Integer.MAX_VALUE;
-		int btm = Integer.MIN_VALUE;// Integer.MIN_VALUE;
-		if (top > (int) tri.v[0].p.getY()) {
-			top = (int) tri.v[0].p.getY();
-		}
-		if (top > (int) tri.v[1].p.getY()) {
-			top = (int) tri.v[1].p.getY();
-		}
-		if (top > (int) tri.v[2].p.getY()) {
-			top = (int) tri.v[2].p.getY();
-		}
-		if (btm < (int) tri.v[0].p.getY()) {
-			btm = (int) tri.v[0].p.getY();
-		}
-		if (btm < (int) tri.v[1].p.getY()) {
-			btm = (int) tri.v[1].p.getY();
-		}
-		if (btm < (int) tri.v[2].p.getY()) {
-			btm = (int) tri.v[2].p.getY();
-		}
-		if (top < 0) {
-			top = 0;
-		}
-		if (btm > BUFFERH) {
-			btm = BUFFERH;
-		}
-
-		// Maximum and minimum buffer initialization
-		for (int i = top; i < btm; i++) {
-			min[i] = Integer.MAX_VALUE;
-			max[i] = Integer.MIN_VALUE;
-		}
-
-		ScanEdge(tri.v[0], tri.v[1]);
-		ScanEdge(tri.v[1], tri.v[2]);
-		ScanEdge(tri.v[2], tri.v[0]);
-
-		// To be drawn on the basis of the maximum and minimum buffer.
-		for (int y = top; y < btm; y++) {
-
-			// Skip if the buffer is not updated
-			if (min[y] == Integer.MAX_VALUE) {
-				continue;
-			}
-
-			int offset = y * BUFFERW;
-
-			// Increment calculation
-			int l = (max[y] - min[y]) + 1;
-			int addr = (maxr[y] - minr[y]) / l;
-			int addg = (maxg[y] - ming[y]) / l;
-			int addb = (maxb[y] - minb[y]) / l;
-			double addu = (maxu[y] - minu[y]) / l;
-			double addv = (maxv[y] - minv[y]) / l;
-
-			int r = minr[y];
-			int g = ming[y];
-			int b = minb[y];
-			double u = minu[y];
-			double v = minv[y];
-
-			for (int x = min[y]; x <= max[y]; x++, r += addr, g += addg, b += addb, u += addu, v += addv) {
-
-				if (x < 0 || x >= BUFFERW) {
-					continue;
-				}
-
-				// flattened pixel index
-				int p = offset + x;
-
-				byte renderFace = isFaceOrderFlipped() ? OverlapRelationValues.UPPER
-						: OverlapRelationValues.LOWER;
-
-				if (zbuf[p] == -1 || interpolatedOverlapRelation.get(zbuf[p], id) == renderFace) {
-
-					int tr = r >> 16;
-					int tg = g >> 16;
-					int tb = b >> 16;
-
-					if (!fillFaces) {
-						pbuf[p] = 0xffffffff;
-
-					} else {
-						if (bUseTexture) {
-							int tx = (int) (textureImage.getWidth() * u);
-							int ty = (int) (textureImage.getHeight() * v);
-
-							tx = tx % textureImage.getWidth();
-							ty = ty % textureImage.getHeight();
-							int textureColor = textureImage.getRGB(tx, ty);
-
-							if (fillFaces && (tri.isFaceFront() ^ isFaceOrderFlipped())) {
-								pbuf[p] = textureColor;
-							} else {
-								pbuf[p] = (tr << 16) | (tg << 8) | tb | 0xff000000;
-
-							}
-						} else {
-							pbuf[p] = (tr << 16) | (tg << 8) | tb | 0xff000000;
-						}
-					}
-					zbuf[p] = id;
-				}
-			}
-		}
-	}
-
-	// --------------------------------------------------------------------
-	// ScanEdge
-	//
-	// Vector v1 ...Starting point
-	// Vector v2 ...Starting point
-	// --------------------------------------------------------------------
-	private void ScanEdge(final TriangleVertex v1, final TriangleVertex v2) {
-
-		int l = Math.abs((int) (v2.p.getY() - v1.p.getY())) + 1;
-
-		// Increment calculation
-		int addx = (int) ((v2.p.getX() - v1.p.getX()) * 0xffff) / l;
-		int addy = (int) ((v2.p.getY() - v1.p.getY()) * 0xffff) / l;
-
-		int addr = (int) (255 * 0xffff * (v2.color.getR() - v1.color.getR()) / l);
-		int addg = (int) (255 * 0xffff * (v2.color.getG() - v1.color.getG()) / l);
-		int addb = (int) (255 * 0xffff * (v2.color.getB() - v1.color.getB()) / l);
-
-		double addu = (v2.uv.getX() - v1.uv.getX()) / l;
-		double addv = (v2.uv.getY() - v1.uv.getY()) / l;
-
-		// Initial value setting
-		int x = (int) (v1.p.getX() * 0xffff);
-		int y = (int) (v1.p.getY() * 0xffff);
-		int r = (int) (255 * 0xffff * v1.color.getR());
-		int g = (int) (255 * 0xffff * v1.color.getG());
-		int b = (int) (255 * 0xffff * v1.color.getB());
-		double u = v1.uv.getX();
-		double v = v1.uv.getY();
-
-		// Scan
-		for (int i = 0; i < l; i++, x += addx, y += addy, r += addr, g += addg, b += addb, u += addu, v += addv) {
-			int py = y >> 16;
-			int px = x >> 16;
-
-			if (py < 0 || py >= BUFFERH) {
-				continue;
-			}
-
-			if (min[py] > px) {
-				min[py] = px;
-				minr[py] = r;
-				ming[py] = g;
-				minb[py] = b;
-				minu[py] = u;
-				minv[py] = v;
-			}
-
-			if (max[py] < px) {
-				max[py] = px;
-				maxr[py] = r;
-				maxg[py] = g;
-				maxb[py] = b;
-				maxu[py] = u;
-				maxv[py] = v;
-			}
-		}
-	}
-
-	private void drawEdges() {
-		// apply Sobel filter
-		for (int y = 1; y < BUFFERH - 1; y++) {
-			for (int x = 1; x < BUFFERW - 1; x++) {
-				int val_h = -1 * zbuf[getIndex(x - 1, y - 1)]
-						+ zbuf[getIndex(x + 1, y - 1)]
-						+ -2 * zbuf[getIndex(x - 1, y)]
-						+ 2 * zbuf[getIndex(x + 1, y)]
-						+ -1 * zbuf[getIndex(x - 1, y + 1)]
-						+ zbuf[getIndex(x + 1, y + 1)];
-				int val_v = -1 * zbuf[getIndex(x - 1, y - 1)]
-						+ zbuf[getIndex(x - 1, y + 1)]
-						+ -2 * zbuf[getIndex(x, y - 1)]
-						+ 2 * zbuf[getIndex(x, y + 1)]
-						+ -1 * zbuf[getIndex(x + 1, y - 1)]
-						+ zbuf[getIndex(x + 1, y + 1)];
-
-				if (val_h != 0 || val_v != 0) {
-					pbuf[getIndex(x, y)] = 0xff888888;
-				}
-			}
-		}
-	}
-
-	private void applyAmbientOcculusion() {
-		byte renderFace = isFaceOrderFlipped() ? OverlapRelationValues.UPPER
-				: OverlapRelationValues.LOWER;
-		int r = 10;
-		int s = (int) (r * r * Math.PI);
-		// For every pixel
-		for (int y = 1; y < BUFFERH - 1; y++) {
-			for (int x = 1; x < BUFFERW - 1; x++) {
-				int f_id = zbuf[getIndex(x, y)];
-
-				// Within a circle of radius r, Count the pixels of the
-				// surface
-				// that is above their own
-				int cnt = 0;
-				for (int dy = -r; dy <= r; dy++) {
-					for (int dx = -r; dx <= r; dx++) {
-						if (dx * dx + dy * dy > r * r) {
-							continue;
-						}
-						if (y + dy < 0 || y + dy > BUFFERH - 1) {
-							continue;
-						}
-						if (x + dx < 0 || x + dx > BUFFERW - 1) {
-							continue;
-						}
-						int f_id2 = zbuf[getIndex(x + dx, y + dy)];
-
-						if (f_id == -1 && f_id2 != -1) {
-							cnt++;
-						} else {
-							if (f_id2 != -1 && interpolatedOverlapRelation.get(f_id, f_id2) == renderFace) {
-								cnt++;
-							}
-						}
-					}
-				}
-
-				if (cnt > 0) {
-					int prev = pbuf[getIndex(x, y)];
-					double ratio = 1.0 - ((double) cnt) / s;
-					int p_r = (int) Math.max(0, ((prev & 0x00ff0000) >> 16) * ratio);
-					int p_g = (int) Math.max(0, ((prev & 0x0000ff00) >> 8) * ratio);
-					int p_b = (int) Math.max(0, (prev & 0x000000ff) * ratio);
-
-					pbuf[getIndex(x, y)] = (p_r << 16) | (p_g << 8) | p_b | 0xff000000;
-				}
-			}
-		}
+				pixelRenderer.width / (domain.getWidth()),
+				pixelRenderer.height / (domain.getHeight())) * 0.95;
 	}
 
 	public void setScale(final double newScale) {
@@ -808,14 +366,6 @@ public class FoldedModelScreen extends JPanel
 		scale *= scale_;
 		updateAffineTransform();
 		repaint();
-	}
-
-	public AffineTransform getAffineTransform() {
-		return affineTransform;
-	}
-
-	public BufferedImage getBufferImage() {
-		return bufferImage;
 	}
 
 	public boolean isFaceOrderFlipped() {
